refactor(eventTicketLogic): extract ticket rules into a pure logic module
